+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## v1.3.1
 
-- Changed `defaultOptions` to `defaultTrackOptions` in the hglib.viewer options.
-- Clear mouse position indicator when leaving the root dom element, on blurring the window, and when no track is hovered
+- Add `constIndicators` as an option to 1D tracks for displaying a visual indicator (e.g., a line) at a constant value.
 
 ## v1.3.0
 
@@ -17,6 +16,10 @@
 - Add a fudge factor to ensure that the entire view is shown in the grid layout
 - Fix minified build
 - Fix a minor visual glitch in the gene annotation track
+- Expose `mix()` from `mixwith` to plugin tracks
+- Support different aggregation modes (mean, min, max, std) for bigWig files
+- Changed `defaultOptions` to `defaultTrackOptions` in the hglib.viewer options.
+- Clear mouse position indicator when leaving the root dom element, on blurring the window, and when no track is hovered
 
 ## v1.2.8